// kodikosERP-master/src/app/models/formatki/sekretariatform.model.ts
export interface Sekretariatform {
    comments?: boolean;
    status?: boolean;
    buttonDodajKontr?: boolean;
    buttonOfertaKontr?: boolean;
    buttonEdytujKontr?: boolean;
    buttonUsunKontr?: boolean;
    projekty?: boolean;
    firmy?: boolean;
    widziWszystkie?: boolean;
    pliki?: boolean;
    buttonPliki?: boolean;
  }