// kodikosERP-master/src/app/models/formatki/kluczewydaniaform.model.ts
export interface KluczeWydaniaform {
  menuKluczeUprawnienia?: boolean;
  menuDodajUprawnienia?: boolean;
  dodajUprSaAktywne?: boolean;
  menuSpis?: boolean;
  menuDodajSpis?: boolean;
  menuWydania?: boolean;
  menuDodajWydanie?: boolean;
  rejestracja?: boolean;
  comments?: boolean;
  status?: boolean;
  buttonDodajKontr?: boolean;
  buttonOfertaKontr?: boolean;
  buttonEdytujKontr?: boolean;
  buttonUsunKontr?: boolean;
  widziWszystkie?: boolean;
}