// kodikosERP-master/src/app/models/formatki/przesylkiform.model.ts
export interface Przesylkiform {
  comments?: boolean;
  status?: boolean;
  buttonDodajKontr?: boolean;
  buttonOfertaKontr?: boolean;
  buttonEdytujKontr?: boolean;
  buttonUsunKontr?: boolean;
  projekty?: boolean;
  firmy?: boolean;
  widziWszystkie?: boolean;
  pliki?: boolean;
  buttonPliki?: boolean;
}